Changes to chiseled block inventory rendering, drawn region & bugfixes
- Added drawn region mode
- Breaking chiseled blocks now drops a chiseled block item.
- Holding chiseled blocks now shows a transparent placement ghost.
- Selecting item modes re-showed the item name in the actionbar instead of as the highlight name.
- Fixed issue where chiseled block items didn't render.
- Fixed issue where chiseled block items didn't have proper block colours.
- Fixed issue where pick block crashed the game and corrupted the world.
- Fixed issue where textures to render couldn't be fetched.
- Fixed issue where unbreakable blocks could be chiseled.
- Fixed issue where breaking chiseled blocks in creative didn't create a sound or made an animation occur.

+++ src/main/java/nl/dgoossens/chiselsandbits2/common/chiseledblock/voxel/IntegerBox.java
@@ -6,7 +6,7 @@ import net.minecraft.util.math.AxisAlignedBB;
 public final class IntegerBox {
     public static final IntegerBox NULL = new IntegerBox(0, 0, 0, 0, 0, 0);
 
-    public double minX, minY, minZ, maxX, maxY, maxZ;
+    public int minX, minY, minZ, maxX, maxY, maxZ;
 
     public IntegerBox(final int x1, final int y1, final int z1, final int x2, final int y2, final int z2) {
         minX = x1;
